Got rid of bracket syntax for arrays. Frees [] up for use with generics.
This is a surprisingly deep change. In addition to getting rid of ArrayExpr and everything related to it, it means arrays are now callables. This exposed some ugly cases where callables didn't work properly when used as fields.

Cleaned up AssignExpr to just handle the basic a = b case and not the stuff that desugars to a message anyway. Eventually, should get rid of it completely.

Also added design question docs about this and generics.