Integrated missing members more fully into the type system.
By definition, a missing member now returns "nothing" without generating a runtime error. This means getMemberType will always return a valid Type because there's no need to distinguish between a missing member and a defined one whose type is Nothing. In other words, every object has an infinite number of members, most of whose type is Nothing.

This will make subtyping records and gets the static checking to be closer to what the dynamic behavior actually is.

Also modified OrType to have the type of a member be the union of its cases. Mirrors the runtime behavior more closely.