Make Callable#getType() return a live type Obj instead of a FunctionType.
Also moved 'isStatic' from Function to FunctionType since it logically
describes the type of a function. This paves the way for allowing fields to
not need an explicit type annotation if they have an initializer. (To support
that, FieldGetter and -Setter will return a type Obj calculated from the
initializer, which will be a live Obj and not a FunctionType.)