Tried to clean up how getter types are handled a bit.
Before, we would evaluate an entire function type for a getter and then pull
out the return type and discard the rest, which is kind of pointless. Now, it
(for the most part) just treats the type of a getter as its value type
directly.

Still need to apply similar clean-up to setters, and it's still a bit fishy
because defining a user-defined getter really does just put the value type as
the return type of the body function, but it's a little better.