Add commas to record syntax: "x: 1, y: 2".
Records and tuples now exist at the same precedence level. This means that
cannot be mixed without using parentheses. For example:

  x: 1, y: 2 // ok
  x: 1, 2    // error!
  x: (1, 2)  // ok

This hopefully makes things a little clearer when you have records like:

  apple: banana cherry: date

Where it looks like "banana cherry" are an expression together. This also makes
multi-line records nicer:

  apple: banana,
  cherry: date

Also rolled ":" directly into the lexer. Names that end with ":" are different
token types. This means that where before "foo   :" would be valid, it now is
not. The ":" must be directly after the name.