Turn operators into function calls.
Before, operators would desugar to a message dispatched on the left-hand 
argument. With this change, they instead become bare function calls with
the arguments passed in as a tuple. This allows more freedom in defining
operator semantics, and gets rid of some dumb code where you have to add 
the same operator method to every class that can be on the left-hand 
side.

Lots of the base library had to be refactored for this to work. It's not
way slower too, for some reason. :(

To avoid ambiguity, ~ is now used for string concatenation.

Added support to "def" for defining functions with operator names.

Added a design doc talking about this a bit more.