Parse "return" expressions in Magpie.
Another step towards fully moving parsewords into Magpie. This required
reorganizing the base lib heavily so that it can bootstrap itself in the
right order. Getting that right is finicky.

Still have lots of work to do getting everything in Magpie. Once that's
done, there's going to be a lot of duplication, and a lot of marshalling
between Magpie and Java. It will be slow.

However, when we go to bytecode compiling, it should get lots faster: 
We'll be able to take the AST code *completely* out of Java. This will
require implementing the entire parser in Magpie, but will get rid of
all of the marshalling back and forth. There will just be one 
translation step: sending a chunk of bytecode from Magpie to Java.