Move parsing "get" expressions into Magpie.
This just covers the standalone ones, not getters within classes and
interfaces.

Got rid of the SpecialFormDesugarer class. I had some bugs where somehow
special forms apply expressions were slipping through and getting checked
instead of being desugared first. I figured the safest way to ensure that
doesn't happen is to have ApplyExpr itself do the desugaring on 
construction. That way, it's impossible to have an Expr that is a special
form apply. As a nice side effect, the code got smaller.