Prep work for parsing "if" in Magpie.
Added support for non-parseword keywords, which is what "then" and "else"
will be: they need to not be recognized as regular identifiers, but they
don't have their own parser.

Refactored parsers to use %if% special form.

Added a %block% special form for blocks. This part is particularly nasty. It may be worth changing the overall Magpie syntax so that blocks 
are part of the core syntax and not just managed by different keywords.
I'll have to play around and see if I can get something I like.