Clean up block grammar.
This required changing the "match" syntax very slightly (and "end" for the
entire match is always required), but makes the process of parsing a block
much simpler and cleaner. The BlockOptions stuff is gone, and the inputs
and outputs to parseBlock() are straightforward, which will make it easier
to expose that to Magpie for the parseword stuff.