Added support to directly bind from a pattern.
The original implementation of "match" would desugar a pattern's variable
bindings to a bunch of "var" expressions so that the interpreter didn't
have to be directly aware of patterns. When I started using patterns for
functions, it did that too, against a magical __args__ variable. Lame.

This gets rid of that hack.

The next step will be to make pattern binding directly supported in the
interpreter and then the desugaring can go away completely.