Make "match" desugar to "var" for binding.
A match expression used to desugar to a nasty mess of variable 
declarations and destructuring traversal for each variable in the matched
pattern. Now that destructuring has first class support, all of that can
be ditched. Yay!

To make match expressions fully first class, the last remaining piece is
getting rid of createPredicate and implementing direct support for pattern
testing.