declareField uses quote for type instead of fn.
This minor change required moving around a bunch of the AST conversion code. The
core AST classes are now defined in Java so that they're immediately available
before any code is parsed. This way, core syntax that uses quotations can be
converted.

I did this so that declareField and defineField can take a quotation for the type
annotation instead of a function. That's the right format for an annotation since
it just needs to be an expression as a chunk of data, and doesn't need to keep a
closure. That's what quotations are for.