Don't always process unquotes.
Unquotes are evaluated during the process of translating an Expr
from Java to Magpie, called by ExprEvaluator. That's the correct
behavior, but it isn't correct for *every* place that translates
from Java to Magpie to do that. When an Expr is just being handed
over to Magpie for a parser, it should be translated over as-is.
This fixes that.