Define actual operator parsers for all operators.
This is the first phase of getting rid of built-in support for any
kind of infix operators or operator identifiers. Now that custom
parsers are defined for each actual operator, I can pull out the old
stuff. Then I can spread out the precedence levels of the new
operators so that operators actually work like they do in other
languages.