Refactor the core class/mixin graph.
The metaclass for Foo is now FooMetaclass. There is now a separate
metaclass for Class (ClassMetaclass) which is distinct from the root
metaclass (just Metaclass). This way, Class new() can be defined in
ClassMetaclass without it also rolling into all of the other 
metaclasses. I *think* these are finally set up right now.