Clean up AST to reflect multimethods.
Instead of a monolithic MessageExpr that handles both methods and
variables (and has a name that doesn't make sense), there is now
CallExpr for method calls and VariableExpr for variable lookup.

It's still a little fishy, though, because a VariableExpr may be
a method call using implicit "this". Still need to figure that out.