Get rid of ugly "->" syntax.
Methods now require blocks. The newline is used to tell when a
getter's pattern has ended and the body begun.

This means you can't do single-line methods anymore, but that's
not much of a loss. I'm leaning towards Python-style function
comments inside the body of the function, so no documented method
can be a single-liner anyway.