Starting to hack doc comments in.
Line comments starting with /// are now special "doc" comments.
They will survive the lexing process and make it all the way to the
parser (which means they are only allowed where the parser expects
them).

Currently, that means at the top of a method body, and the top of
a class declaration. If present there, the parser will grab them
and add them to the method or class definition.

Method docs can be accessed at runtime using:

    docMethod("methodName")

Class docs can be access using:

    SomeClass doc