Linearize records with differing fields.
Can now correctly linearize records where one is a subset of the
other. For example, given:

def foo(a is Int, b is String, c is Bool) ...
def foo(a is Int, b is String) ...

Then this:

foo(1, "2", true)

Matches both. The linearizer can now handle this (and prefers the
former since it captures more of the argument).