Reorganize spec to not rely on import order.
specify() now just adds the test to a list of tests that get run
when runTests() is called. This way you don't have to interleave
the spec imports with setting up the tests.

This also means that time to parse the specs is no longer included
in the reported time.