Tweak method desugaring (again).
Before, calls desugared like this:
a b    ->  b(a)
b(c)   ->  b(nothing, c)
a b(c) ->  b(a, c)

The problem is that that means these desugar to the same thing:

(a, c) b
a b(c)

Which is weird. It means you could do:

([1, 2, 3], 4) add

Instead of 

[1, 2, 3] add(4)

With this change, a call always desugars to a two field record, so
you can't do that. It's a little simpler this way. A getter is
basically an inferred-nothing right arg, and a function-like
method is an inferred-nothing left arg.