Add join() and toString for collections.
There is now an Iterable base class that anything with an iterate()
method should inherit. join() is specialized to it.

Implemented (and spec'd) toString for Array and List. Also filled
in spec for other existing types.

Added specifyMethod(). Eventually most specs should move to using
that instead of specify to get rid of the "The __ method on __."
boilerplate.