Assignment expressions.
That was surprisingly easy (<3 patterns!). Simple assignment,
multiple assignment, destructuring records, etc. are all working.

Correctly reports a compile error if you assign to an undeclared
variable.

The big missing piece is correctly reporting an error if you assign
to a "val" variable, which should be easy to support. (The resolver
just needs to keep track of which locals are immutable.)