Get while loops working.
After some thought, I've decided to not unify "while" and "for" into
a single loop construct that allows multiple clauses. I found a chain
of "while" or "for" without "do" (i.e. a parallel loop) to be too
easy to confuse with *nested* loops.

The main use case to allow this was parallel iteration through 
collections. For that, a much simpler solution now that we have
destructuring, is to just implement "iterate" on tuple types so
you can do:

  for a, b in [1, 2, 3], [2, 3, 4] do ...